Safari-609.1.11.1/LayoutTests/http/tests/cache/disk-cache/disk-cache-redirect-to-data.html

<script src="/js-test-resources/js-test-pre.js"></script>
<script src=resources/redirect-to-data.php></script>
<script>
description("Test that HTTP redirect to a data URL does not crash or assert in the network process.");
shouldBeTrue("success");
</script>
<script src="/js-test-resources/js-test-post.js"></script>
